fix/QA-116: 9,999,999,999원 입력 시 오류 발생

+++ core/model/src/main/java/com/susu/core/model/FriendStatistics.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.Stable
 @Stable
 data class FriendStatistics(
     val friend: Friend = Friend(),
-    val receivedAmounts: Int = 0,
-    val sentAmounts: Int = 0,
-    val totalAmounts: Int = 0,
+    val receivedAmounts: Long = 0,
+    val sentAmounts: Long = 0,
+    val totalAmounts: Long = 0,
 )
